refactor(usd): split types.rs into per-type sibling modules
types.rs (847 lines) becomes a 70-line thin facade that mounts six per-type siblings via #[path]: types_sdf_path (SdfPath), types_vt_value (VtValue), types_attribute (UsdAttribute), types_prim (UsdPrim + default_true serde helper), types_layer (UsdLayer + default_y_axis / default_mpu serde helpers), and types_metrics (UsdStageMetrics). Unit tests move into types_tests.rs.

Every public type is re-exported so dcc_mcp_usd::types::{SdfPath, VtValue, UsdAttribute, UsdPrim, UsdLayer, UsdStageMetrics} keeps working unchanged for downstream callers.

93 lib tests + 3 doctests green; docs/guide/architecture.md gains a Maintainer layout section under dcc-mcp-usd.
